Move code to 'features/products.js'
