+++ src/redux/reducer/index.js
@@ -1,6 +1,6 @@
 import order from '../features/order';
 import restaurants from '../features/restaurants';
-import products from './products';
+import products from '../features/products';
 import reviews from '../features/reviews';
 import users from './users';
