+++ src/redux/selectors.js
@@ -1,23 +1,16 @@
 import { createSelector } from '@reduxjs/toolkit';
 import { orderSelector } from './features/order';
+import { productsSelector } from './features/products';
 import { restaurantSelector } from './features/restaurants';
 import { reviewSelector, reviewsSelector } from './features/reviews';
-import { isLoaded, isLoading, shouldLoad } from './utils';
+import { isLoaded, shouldLoad } from './utils';
 
-const productsSelector = (state) => state.products.entities;
 const usersSelector = (state) => state.users.entities;
 
-const productsStatusSelector = (state, props) =>
-  state.products.status[props.restaurantId];
 const usersStatusSelector = (state) => state.users.status;
-
-export const productsLoadingSelector = isLoading(productsStatusSelector);
-export const shouldLoadProductsSelector = shouldLoad(productsStatusSelector);
 
 export const usersLoadedSelector = isLoaded(usersStatusSelector);
 export const shouldLoadUsersSelector = shouldLoad(usersStatusSelector);
-
-export const productSelector = (state, { id }) => productsSelector(state)[id];
 
 export const orderProductsSelector = createSelector(
   productsSelector,
